# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Serverless CMF with Full Rest API and RBAC(Role Base Control Access) System.",
        "title": "MayCMF",
        "contact": {
            "name": "eneus"
        },
        "license": {},
        "version": "0.1.0"
    },
    "host": "127.0.0.1:8088",
    "basePath": "/",
    "paths": {
...
        "/api/v1/language": {
            "get": {
                "tags": [
                    "Language"
                ],
                "summary": "Query data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page Index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Paging Size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Numbering",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Status (1: Enable 2: Disable)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search result: {list:List data,pagination:{current:Page index, pageSize: Page size, total: The total number}}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Language"
                            }
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Language"
                ],
                "summary": "Create data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Create data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/schema.Language"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Language"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message: Invalid request parameter}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/language/{code}": {
            "get": {
                "tags": [
                    "Language"
                ],
                "summary": "Query specified data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Language"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "404": {
                        "description": "{error:{code:0,message: Resource does not exist.}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Language"
                ],
                "summary": "Update data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/schema.Language"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Language"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message: Invalid request parameter}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Language"
                ],
                "summary": "Delete data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/language/{code}/disable": {
            "patch": {
                "tags": [
                    "Language"
                ],
                "summary": "Disable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/language/{code}/enable": {
            "patch": {
                "tags": [
                    "Language"
                ],
                "summary": "Enable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error item",
                    "type": "object",
                    "$ref": "#/definitions/schema.HTTPErrorItem"
                }
            }
        },
        "schema.HTTPErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "schema.HTTPStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "status(OK)",
                    "type": "string"
                }
            }
        },
        "schema.Language": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "active": {
                    "description": "Status language can be added as content",
                    "type": "boolean"
                },
                "code": {
                    "description": "Code",
                    "type": "string"
                },
                "default": {
                    "description": "Default language for project",
                    "type": "boolean"
                },
                "name": {
                    "description": "Name",
                    "type": "string"
                },
                "native": {
                    "description": "Remarks",
                    "type": "string"
                },
                "rtl": {
                    "description": "RTL",
                    "type": "boolean"
                }
            }
        },
...
    }
}